Reunifying Tests and Basic Skeleton for wpDeletePost
